third_party/WebKit/LayoutTests/fast/html/imports/import-style-tree-order-dedup-expected.txt

PASS colorOf(target1) is 'rgb(0, 128, 0)'
PASS colorOf(target2) is 'rgb(255, 0, 0)'
PASS colorOf(target3) is 'rgb(255, 0, 0)'
PASS colorOf(target1) is 'rgb(0, 128, 0)'
PASS colorOf(target2) is 'rgb(0, 0, 255)'
PASS colorOf(target3) is 'rgb(255, 0, 0)'
PASS colorOf(target1) is 'rgb(0, 128, 0)'
PASS colorOf(target2) is 'rgb(0, 0, 255)'
PASS colorOf(target3) is 'rgb(255, 0, 0)'
PASS successfullyParsed is true

TEST COMPLETE
Hello! Hello! Hello!
